+++ CHANGELOG.md
@@ -10,6 +10,15 @@ Official page: [http://kiwigrid.github.io/keta/](http://kiwigrid.github.io/keta/
 
 # Changelog
 
+## Version 0.4.5
+
+_Released: 2015-08-17_
+
+* Bugfix for `ExtendedTable` directive: non-heterogeneous rows don’t break table anymore
+* Added `ApplicationUtils.getAppIcon` to retrieve app icon for specified language
+* Added `ApplicationUtils.getAppAuthor` to retrieve app author
+* Updated  css parts
+
 ## Version 0.4.4
 
 _Released: 2015-07-30_